CKFK 20220110 Correcciones con las fechas

+++ app/src/main/java/com/dts/roadp/clsDocCanastaBod.java
@@ -61,7 +61,7 @@ public class clsDocCanastaBod extends clsDocument {
             ref=DT.getString(3);
             vend=DT.getString(4);
 
-            ffecha=DT.getInt(5);
+            ffecha=DT.getLong(5);
             fsfecha=sfecha(ffecha);
 
             corel=serie;
